chore: shifting the topic types to enum
