+++ packages/common/src/types.ts
@@ -13,7 +13,7 @@ import {
     networkPublicKey,
 } from "./validations";
 import {Message} from "@solana/web3.js";
-import {BitcoinCluster, EthCluster, SolCluster, USDCCluster} from "./constant";
+import {BitcoinCluster, EthCluster, SolCluster, USDCCluster} from "./enum";
 import {Cluster} from "@solana/web3.js";
 import { BitcoinChainId, BitcoinToken, EthChainId, EthToken, MsgTopics, Network, NotifTopics, SolChainId, SolToken, TopicTypes, WsMessageTypeEnum, hookStatus } from "./enum";
 import { BtcExplorer, EthExplorer, SolExplorer } from "./settings";
